Special utf-8 detection in read_auto().
After trying different UTF BOMs, a valid utf-8 footprint is
checked in the first bunch of characters in the file. If it's not
valid utf-8, encoding is set back to iso8859-1. This is not
really correct; it can be koi-8 or another oddity. The right thing
would be to *set* encoding to utf-8 if explicitly detected and fall
back to locale encoding otherwise, but this will fail again on
utf-8 environments.